update Blob interface, move interfaces
