+++ jso/apis/src/main/java/org/teavm/jso/dom/html/HTMLInputElement.java
@@ -16,7 +16,7 @@
 package org.teavm.jso.dom.html;
 
 import org.teavm.jso.JSProperty;
-import org.teavm.jso.browser.FileList;
+import org.teavm.jso.file.FileList;
 
 /**
  *
