+++ jso/apis/src/main/java/org/teavm/jso/typedarrays/ArrayBufferView.java
@@ -17,12 +17,13 @@ package org.teavm.jso.typedarrays;
 
 import org.teavm.jso.JSObject;
 import org.teavm.jso.JSProperty;
+import org.teavm.jso.blob.BlobConvertible;
 
 /**
  *
  * @author Alexey Andreev
  */
-public abstract class ArrayBufferView implements JSObject {
+public abstract class ArrayBufferView implements JSObject, BlobConvertible {
     protected ArrayBufferView() {
     }
